requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-ckeditor==6.3.2
django-image-optimizer==1.0.1
django-js-asset==2.0.0
django-nested-admin==3.4.0
django-optimized-image==0.3.0
django-resized==0.3.11
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.3.4
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-monkey-business==1.0.0
python-resize-image==1.1.20
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tinify==1.5.2
toposort==1.7
urllib3==1.26.9
watchdog==2.1.2
whitenoise==6.0.0
